move observations from upload to meta :)
